# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API token authentication. Use format: \"Bearer \u003cyour-api-token\u003e\"",
        "title": "Bible Notes API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "shuvoedward@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/v1",
    "paths": {
...
        "/v1/notes": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a paginated list of notes filtered by type (GENERAL or BIBLE) with optional sorting",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "List user notes",
                "parameters": [
                    {
                        "enum": [
                            "GENERAL",
                            "BIBLE"
                        ],
                        "type": "string",
                        "description": "Note type",
                        "name": "note_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 10000,
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "Number of items per page",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "created_at",
                            "-created_at",
                            "title",
                            "-title"
                        ],
                        "type": "string",
                        "default": "-created_at",
                        "description": "Sort field and direction",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved notes (key is note_type)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/shuvoedward_Bible_project_internal_data.NoteMetadata"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid query parameters or validation errors",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a new note of type GENERAL, BIBLE, or CROSS_REFERENCE. GENERAL notes only require title and content. BIBLE and CROSS_REFERENCE notes require additional location fields (book, chapter, verses, and offsets), title optional for BIBLE and no title necessary for CROSS_REFERENCE.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "Create a new note",
                "parameters": [
                    {
                        "description": "Note creation data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cmd_api.CreateNoteInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created note",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/shuvoedward_Bible_project_internal_data.NoteResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "Duplicate title (GENERAL), location already linked (BIBLE/CROSS_REFERENCE), or duplicate content",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "object",
                                "additionalProperties": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/notes/{id}/locations": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a location link for an existing note at the specified Bible verse and word offsets",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "Link note to verse location",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Note ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Location details",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully linked note to location",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Note not found or doesn't belong to user",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "cmd_api.CreateNoteInput": {
            "type": "object",
            "properties": {
                "book": {
                    "description": "Bible verse location fields (only used for BIBLE and CROSS_REFERENCE types)",
                    "type": "string"
                },
                "chapter": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "end_offset": {
                    "description": "Character offset within end verse",
                    "type": "integer"
                },
                "end_verse": {
                    "description": "end verse must be provided, included. when just one verse, svs = 1 and evs = 1.",
                    "type": "integer"
                },
                "note_type": {
                    "description": "\"GENERAL\", \"BIBLE\", or \"CROSS_REFERENCE\"",
                    "type": "string"
                },
                "start_offset": {
                    "description": "Character offset within start verse",
                    "type": "integer"
                },
                "start_verse": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "shuvoedward_Bible_project_internal_data.NoteMetadata": {
            "type": "object",
            "properties": {
                "Preview": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "note_type": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "shuvoedward_Bible_project_internal_data.NoteResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "location": {
                    "$ref": "#/definitions/shuvoedward_Bible_project_internal_data.LocationResponse"
                },
                "note_id": {
                    "type": "integer"
                },
                "note_type": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "security": [
        {
            "ApiKeyAuth": []
        }
    ]
}